update titles and file names

+++ graph_utility/best_overall.py
@@ -10,7 +10,7 @@ import utility
 
 
 def get_strictly_better_points(derived_jable, metric, test_names, keep_largest_percent, how_much_better):
-    derived_jable = utility.largest_x(derived_jable, keep_largest_percent, metric, test_names)
+    derived_jable = utility.largest_x_jable(derived_jable, keep_largest_percent, metric, test_names)
 
     metric_columns = [experiment_column + '@' + metric for experiment_column in test_names]
 
@@ -44,7 +44,7 @@ def get_strictly_better_points(derived_jable, metric, test_names, keep_largest_p
 
 
 def get_eq_points(derived_jable, metric, test_names, keep_largest_percent):
-    derived_jable = utility.largest_x(derived_jable, keep_largest_percent, metric, test_names)
+    derived_jable = utility.largest_x_jable(derived_jable, keep_largest_percent, metric, test_names)
 
     metric_columns = [experiment_column + '@' + metric for experiment_column in test_names]
 
@@ -87,6 +87,7 @@ def get_answer_df(derived_jable, test_names):
 def plot(data_list, test_names, graph_dir, keep_largest_percent, how_much_better):
     data_list = copy.deepcopy(data_list)
     test_names = copy.deepcopy(test_names)
+    png_names = ['all', 'with', 'without']
 
     if len(test_names) == 2 and 'no-red' in test_names:
         print(
@@ -113,25 +114,7 @@ def plot(data_list, test_names, graph_dir, keep_largest_percent, how_much_better
          'unique answers': utility.zero_padding(derived_jable['unique answers'].value_counts(), 'unique answers',
                                                 test_names).tolist(),
          }, index=test_names)
-
-    points_eq_df = pd.DataFrame(
-        {'reduced size': get_eq_points(derived_jable, 'reduced size', test_names, keep_largest_percent
-                                       ),
-         'state space size': get_eq_points(derived_jable, 'state space size', test_names, keep_largest_percent
-                                           ),
-         'reduce time': get_eq_points(derived_jable, 'reduce time', test_names, keep_largest_percent),
-         'verification memory': get_eq_points(derived_jable, 'verification memory', test_names, keep_largest_percent
-                                              ),
-         'verification time': get_eq_points(derived_jable, 'verification time', test_names, keep_largest_percent
-                                            ),
-         'total time': get_eq_points(derived_jable, 'total time', test_names,
-                                     keep_largest_percent),
-         'answered queries': answer_df,
-         }, index=test_names)
-
     data_to_plot = utility.split_into_all_with_without(points_df)
-    data_to_plot_eq = utility.split_into_all_with_without(points_eq_df)
-    png_names = ['all', 'with', 'without']
 
     for index, data in enumerate(data_to_plot):
         if index > 0:
@@ -154,33 +137,53 @@ def plot(data_list, test_names, graph_dir, keep_largest_percent, how_much_better
             plot.annotate(int(width), xy=(left + width, bottom + height / 2), ha='center', va='center', size=10)
 
         plt.savefig(
-            graph_dir + f'{how_much_better * 100}%_better_points_{png_names[index]}_largest_{keep_largest_percent * 100}%tests.png',
+            graph_dir + f'better_than_compare_all_largest_{keep_largest_percent * 100}%_tests_{how_much_better}%_better_{png_names[index]}.png',
             bbox_inches='tight')
         plt.clf()
 
-    for index, data in enumerate(data_to_plot_eq):
-        if index > 0:
-            continue
-        if len(data) == 0:
-            continue
-        # Plot the second plot with with
-        plot = data.plot(kind='barh', width=0.75, linewidth=2, figsize=(10, 10))
-        plt.legend(bbox_to_anchor=(1.02, 1), loc='best', borderaxespad=0)
-        plt.title(
-            f'Point given if at least as good as runner-up, using {keep_largest_percent * 100}% largest tests ({int(derived_jable.shape[0] * keep_largest_percent)} tests)')
-        plt.xscale('log')
-        plt.xlabel("points")
-        plt.ylabel('experiments')
+    if not os.path.isfile(
+            graph_dir + f'eq_compare_all_largest_{keep_largest_percent * 100}%_tests_all.png'):
 
-        # Plot the numbers in the bars
-        for p in plot.patches:
-            left, bottom, width, height = p.get_bbox().bounds
-            plot.annotate(int(width), xy=(left + width, bottom + height / 2), ha='center', va='center', size=10)
+        points_eq_df = pd.DataFrame(
+            {'reduced size': get_eq_points(derived_jable, 'reduced size', test_names, keep_largest_percent
+                                           ),
+             'state space size': get_eq_points(derived_jable, 'state space size', test_names, keep_largest_percent
+                                               ),
+             'reduce time': get_eq_points(derived_jable, 'reduce time', test_names, keep_largest_percent),
+             'verification memory': get_eq_points(derived_jable, 'verification memory', test_names, keep_largest_percent
+                                                  ),
+             'verification time': get_eq_points(derived_jable, 'verification time', test_names, keep_largest_percent
+                                                ),
+             'total time': get_eq_points(derived_jable, 'total time', test_names,
+                                         keep_largest_percent),
+             'answered queries': answer_df,
+             }, index=test_names)
 
-        plt.savefig(
-            graph_dir + f'better_or_eq_points_{png_names[index]}_largest_{keep_largest_percent * 100}%tests.png',
-            bbox_inches='tight')
-        plt.clf()
+        data_to_plot_eq = utility.split_into_all_with_without(points_eq_df)
+
+        for index, data in enumerate(data_to_plot_eq):
+            if index > 0:
+                continue
+            if len(data) == 0:
+                continue
+            # Plot the second plot with with
+            plot = data.plot(kind='barh', width=0.75, linewidth=2, figsize=(10, 10))
+            plt.legend(bbox_to_anchor=(1.02, 1), loc='best', borderaxespad=0)
+            plt.title(
+                f'Point given if at least as good as runner-up, using {keep_largest_percent * 100}% largest tests ({int(derived_jable.shape[0] * keep_largest_percent)} tests)')
+            plt.xscale('log')
+            plt.xlabel("points")
+            plt.ylabel('experiments')
+
+            # Plot the numbers in the bars
+            for p in plot.patches:
+                left, bottom, width, height = p.get_bbox().bounds
+                plot.annotate(int(width), xy=(left + width, bottom + height / 2), ha='center', va='center', size=10)
+
+            plt.savefig(
+                graph_dir + f'eq_compare_all_largest_{keep_largest_percent * 100}%_tests_{png_names[index]}.png',
+                bbox_inches='tight')
+            plt.clf()
 
 
 if __name__ == "__main__":
